fix rtl8192eu driver

+++ lichee/linux-3.4/drivers/net/wireless/rtl8192eu/include/rtw_tdls.h
@@ -134,7 +134,7 @@ void rtw_build_tdls_ch_switch_rsp_ies(_adapter * padapter, struct xmit_frame * p
 void rtw_build_tunneled_probe_req_ies(_adapter * padapter, struct xmit_frame * pxmitframe, u8 *pframe);
 void rtw_build_tunneled_probe_rsp_ies(_adapter * padapter, struct xmit_frame * pxmitframe, u8 *pframe);
 
-int update_sgi_tdls(_adapter *padapter, struct sta_info *psta);
+u8	update_sgi_tdls(_adapter *padapter, struct sta_info *psta);
 u32 update_mask_tdls(_adapter *padapter, struct sta_info *psta);
 #endif //CONFIG_TDLS
